# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AwaitingConfirmationEvent": {
        "properties": {
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SchemaProposalOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AwaitingConfirmationEvent",
        "type": "object"
      },
...
      "ConfirmedColumn": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "semantic_type": {
            "enum": [
              "id",
              "numeric",
              "categorical",
              "binary",
              "datetime",
              "text"
            ],
            "title": "Semantic Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "semantic_type"
        ],
        "title": "ConfirmedColumn",
        "type": "object"
      },
      "ConfirmedSchema": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ConfirmedColumn"
            },
            "title": "Columns",
            "type": "array"
          },
          "id_columns": {
            "items": {
              "type": "string"
            },
            "title": "Id Columns",
            "type": "array"
          },
          "positive_label": {
            "title": "Positive Label",
            "type": "string"
          },
          "revenue_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Column"
          },
          "target_column": {
            "title": "Target Column",
            "type": "string"
          },
          "time_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Column"
          }
        },
        "required": [
          "target_column",
          "positive_label"
        ],
        "title": "ConfirmedSchema",
        "type": "object"
      },
      "DoneEvent": {
        "properties": {
          "final_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "DoneEvent",
        "type": "object"
      },
      "ErrorEvent": {
        "properties": {
          "fatal": {
            "default": false,
            "title": "Fatal",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
      "HTTPErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HTTPErrorOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "NodeFinishEvent": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "status": {
            "enum": [
              "started",
              "done",
              "skipped",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "node",
          "status"
        ],
        "title": "NodeFinishEvent",
        "type": "object"
      },
      "NodeStartEvent": {
        "properties": {
          "node": {
            "title": "Node",
            "type": "string"
          }
        },
        "required": [
          "node"
        ],
        "title": "NodeStartEvent",
        "type": "object"
      },
      "PredictionsPage": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "band": {
            "anyOf": [
              {
                "enum": [
                  "High",
                  "Medium",
                  "Low"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Band"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "available",
          "page",
          "page_size",
          "total",
          "pages",
          "items"
        ],
        "title": "PredictionsPage",
        "type": "object"
      },
      "ProposedColumn": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "semantic_type": {
            "enum": [
              "id",
              "numeric",
              "categorical",
              "binary",
              "datetime",
              "text"
            ],
            "title": "Semantic Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "semantic_type",
          "confidence"
        ],
        "title": "ProposedColumn",
        "type": "object"
      },
      "ResultsResponse": {
        "properties": {
          "predictions": {
            "$ref": "#/components/schemas/PredictionsPage"
          },
          "results": {
            "additionalProperties": true,
            "title": "Results",
            "type": "object"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "awaiting_confirmation",
              "done",
              "failed",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "results",
          "predictions"
        ],
        "title": "ResultsResponse",
        "type": "object"
      },
      "RunStatusResponse": {
        "properties": {
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status"
        ],
        "title": "RunStatusResponse",
        "type": "object"
      },
      "SchemaProblems": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "problems": {
            "items": {
              "type": "string"
            },
            "title": "Problems",
            "type": "array"
          }
        },
        "required": [
          "message",
          "problems"
        ],
        "title": "SchemaProblems",
        "type": "object"
      },
      "SchemaProblemsOut": {
        "description": "Body of a 422 from POST /confirm-schema when the schema does not fit the data.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/SchemaProblems"
          }
        },
        "required": [
          "detail"
        ],
        "title": "SchemaProblemsOut",
        "type": "object"
      },
      "SchemaProposalOut": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ProposedColumn"
            },
            "title": "Columns",
            "type": "array"
          },
          "id_columns": {
            "items": {
              "type": "string"
            },
            "title": "Id Columns",
            "type": "array"
          },
          "positive_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Positive Label"
          },
          "reasoning": {
            "default": "",
            "title": "Reasoning",
            "type": "string"
          },
          "revenue_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Column"
          },
          "source": {
            "enum": [
              "ai+rules",
              "rules"
            ],
            "title": "Source",
            "type": "string"
          },
          "target_candidates": {
            "items": {
              "type": "string"
            },
            "title": "Target Candidates",
            "type": "array"
          },
          "target_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Column"
          },
          "time_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Column"
          }
        },
        "required": [
          "columns",
          "source"
        ],
        "title": "SchemaProposalOut",
        "type": "object"
      },
      "StreamEvents": {
        "description": "Documentation only: the payload type of each SSE event name.",
        "properties": {
          "awaiting_confirmation": {
            "$ref": "#/components/schemas/AwaitingConfirmationEvent"
          },
          "done": {
            "$ref": "#/components/schemas/DoneEvent"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorEvent"
          },
          "heartbeat": {
            "additionalProperties": true,
            "title": "Heartbeat",
            "type": "object"
          },
          "node_finish": {
            "$ref": "#/components/schemas/NodeFinishEvent"
          },
          "node_start": {
            "$ref": "#/components/schemas/NodeStartEvent"
          },
          "resumed": {
            "additionalProperties": true,
            "title": "Resumed",
            "type": "object"
          }
        },
        "required": [
          "node_start",
          "node_finish",
          "error",
          "awaiting_confirmation",
          "resumed",
          "heartbeat",
          "done"
        ],
        "title": "StreamEvents",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ChurnLens API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analyze/{session_id}": {
      "post": {
        "operationId": "analyze_analyze__session_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorOut"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "runs"
        ]
      }
    },
    "/confirm-schema/{session_id}": {
      "post": {
        "operationId": "confirm_schema_confirm_schema__session_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmedSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorOut"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaProblemsOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Confirm Schema",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/results/{session_id}": {
      "get": {
        "operationId": "results_results__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "band",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "High",
                    "Medium",
                    "Low"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Band"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorOut"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Results",
        "tags": [
          "results"
        ]
      }
    },
...
    "/stream/{session_id}": {
      "get": {
        "operationId": "stream_stream__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "description": "Resume after this event id. For clients that cannot send Last-Event-ID, such as a browser EventSource re-created after it closed.",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resume after this event id. For clients that cannot send Last-Event-ID, such as a browser EventSource re-created after it closed.",
              "title": "After"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamEvents"
                }
              },
              "text/event-stream": {}
            },
            "description": "Server-sent events; StreamEvents maps each event name to its data payload."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorOut"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}